tools/Aj Pixel Cut/requirements.txt
flask>=2.2
gunicorn>=21.2
pillow>=10.0
rembg>=2.0
